linkedinreply/requirements.txt
requests==2.25.1
selenium==3.141.0
pymongo==3.11.2
mongoengine==0.22.1
beautifulsoup4==4.9.3
dnspython==2.1.0
